do it & cleanup
