+++ aladdin/config.py
@@ -1,17 +1,16 @@
 import json
 import os
 import pathlib
-import subprocess
 from distutils.util import strtobool
 
 from aladdin import __version__
 from aladdin.lib import logging
-from aladdin.lib.utils import working_directory
 
 logger = logging.getLogger(__name__)
 
 
 PROJECT_ROOT = os.path.dirname(os.path.dirname(__file__))
+ALADDIN_DOCKER_REPO = "fivestarsos/aladdin"
 
 
 def load_cluster_configs():
@@ -51,7 +50,7 @@ def load_config():
     return load_config_from_file(f'{os.environ["ALADDIN_CONFIG_DIR"]}/config.json')
 
 
-def load_user_config_file() -> dict:
+def load_user_config() -> dict:
     home = pathlib.Path.home()
     return load_config_from_file(home / ".aladdin/config/config.json")
 
@@ -62,91 +61,4 @@ def set_user_config_file(config: dict):
         json.dump(config, json_file, indent=2)
 
 
-def set_config_path() -> bool:
-    """
-    Function to set the "ALADDIN_CONFIG_DIR" env var
-    Uses git to fetch the latest revision of the config repo
-
-    NOTE:
-    If this function returns False it will short-circuit execution of
-    any aladdin command, so returning False should be preceded by some
-    user-friendly error statement about why we're exiting
-    """
-    if os.getenv("ALADDIN_CONFIG_DIR"):
-        # Aladdin config is already set, nothing to do here
-        return True
-
-    err_message = (
-        "Unable to find config repo. "
-        "Please use "
-        "'aladdin config set config_repo <git@github.com:{git_account}/{repo}.git>' "
-        "to set config repo"
-    )
-    try:
-        config = load_user_config_file()
-    except FileNotFoundError:
-        logger.error(err_message)
-        return False
-
-    config_dir: str = config.get("config_dir")
-    config_repo: str = config.get("config_repo")
-    if not config_dir and not config_repo:
-        logger.error(err_message)
-        return False
-
-    if not config_repo:
-        # try to auto-set config_repo
-        if not os.path.isdir(config_dir):
-            logger.error(err_message)
-            return False
-        with working_directory(config_dir):
-            try:
-                remote = subprocess.run(
-                    "git remote get-url origin".split(),
-                    check=True,
-                    capture_output=True,
-                    encoding="utf-8"
-                ).stdout.strip()
-            except subprocess.CalledProcessError:
-                logger.error(err_message)
-                return False
-        *_, git_account, repo = remote.split("/")
-        config_repo = f"git@github.com:{git_account}/{repo}"
-        config["config_repo"] = config_repo
-        set_user_config_file(config)
-
-    remote_config_path = pathlib.Path.home() / ".aladdin/remote_config"
-    commands = [f"git clone -b {__version__} {config_repo} remote_config"]
-    cwd = pathlib.Path.home() / ".aladdin"
-    if os.path.isdir(remote_config_path) and os.path.isdir(remote_config_path / ".git"):
-        cwd = remote_config_path
-        commands = [
-            "git fetch --tags --prune -f",
-            f"git checkout {__version__}"
-        ]
-    with working_directory(cwd):
-        for command in commands:
-            try:
-                subprocess.run(
-                    command.split(),
-                    check=True,
-                    encoding="utf-8",
-                    capture_output=True
-                )
-            except subprocess.CalledProcessError as e:
-                logger.error(
-                    "Failed to fetch aladdin config (%s) from remote: \n%s",
-                    config_repo,
-                    e.stderr.strip() or e.stdout.strip()
-                )
-                return False
-
-    os.environ["ALADDIN_CONFIG_DIR"] = str(remote_config_path)
-
-    if strtobool(os.getenv("ALADDIN_DEV", "false")) and config_dir and os.path.isdir(config_dir):
-        """
-        Allow aladdin developers to use a custom config
-        """
-        os.environ["ALADDIN_CONFIG_DIR"] = config_dir
-
-    return True
+ALADDIN_DEV = bool(strtobool(os.getenv("ALADDIN_DEV", "false")))
